BrowserGym/requirements.txt
browsergym-core>=0.3
browsergym-experiments>=0.3
openai>=1.35.4,<1.36
langchain>=0.2,<0.3
langchain_openai>=0.1.10,<0.2
tiktoken
huggingface_hub
contexttimer
ipython
pyyaml>=6
pandas
joblib
transformers
langchain_community>=0.2.6,<0.3
anthropic
httpx==0.27.2 
google-generativeai